+++ pkg/handlers/backup/state_uploaderror.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -27,13 +27,23 @@ import (
 )
 
 func stateUploadErrorHandler(h *handler, backup *backupApi.ArangoBackup) (*backupApi.ArangoBackupStatus, error) {
-	if backup.Spec.Upload == nil || !backup.Status.Backoff.GetNext().After(time.Now()) {
+	// no more retries - move to failed state
+	if !backup.Status.Backoff.ShouldBackoff(backup.Spec.Backoff) {
+		return wrapUpdateStatus(backup,
+			updateStatusState(backupApi.ArangoBackupStateFailed, "out of Upload retries"),
+			cleanStatusJob())
+	}
+
+	// if we should retry - move to ready state
+	if backup.Spec.Upload == nil ||
+		(backup.Status.Backoff.ShouldBackoff(backup.Spec.Backoff) && !backup.Status.Backoff.GetNext().After(time.Now())) {
 		return wrapUpdateStatus(backup,
 			updateStatusState(backupApi.ArangoBackupStateReady, ""),
 			cleanStatusJob(),
 			updateStatusAvailable(true))
 	}
 
+	// no ready to retry - wait (do not change state)
 	return wrapUpdateStatus(backup,
 		updateStatusAvailable(true))
 }